Updated and regactored firebase interactions
