+++ package.json
@@ -31,7 +31,7 @@
     "react-native-bootsplash": "^4.0.2",
     "react-native-bouncy-checkbox": "^2.1.7",
     "react-native-chart-kit": "^6.11.0",
-    "react-native-gesture-handler": "^2.1.0",
+    "react-native-gesture-handler": "^2.2.0",
     "react-native-image-picker": "^4.7.1",
     "react-native-safe-area-context": "^3.3.2",
     "react-native-screens": "^3.10.1",
