--- src/api/lofftApi.ts
@@ -1,5 +1,0 @@
-import axios from "axios";
-
-export default axios.create({
-  baseURL: "http://localhost:3001",
-});
